chore: Add tarpaulin run to ci (#15)
This adds tarpaulin to measure test coverage via codecov.

On the way, I switched to the default rust container as there was a problem installing tarpaulin (openssl-sys crate would not compile, there seems to be an issue with finding openssl on musl). This should not make any difference, we don't build release binaries, only test the library.

Additionally, I integrated some small, new clippy lints to have a fresh green build as well as bumped dependencies for the github actions where necessary.